Migration test complete

+++ e2e/e2e_tests/cluster_test_base.py
@@ -155,10 +155,12 @@ class TestClusterBase:
         except Exception as _:
             self.logger.info(traceback.format_exc())
         try:
-            instance_id = self.common_utils.get_instance_id_by_name(ec2_resource=self.ec2_resource,
-                                                                    instance_name="e2e-new-instance")
-            self.common_utils.terminate_instance(ec2_resource=self.ec2_resource,
-                                                instance_id=instance_id)
+            if self.ec2_resource:
+                instance_id = self.common_utils.get_instance_id_by_name(ec2_resource=self.ec2_resource,
+                                                                        instance_name="e2e-new-instance")
+                if instance_id:
+                    self.common_utils.terminate_instance(ec2_resource=self.ec2_resource,
+                                                         instance_id=instance_id)
         except Exception as e:
             self.logger.info(f"Error while deleting instance: {e}")
             self.logger.info(traceback.format_exc())
